system.exit calls removed, driving (simulation) system test added, fixed bug when out link was nort set for simple Connection, new implementation of getWorldWith and getWorldHeight in PositionUtils - it reflects BoundingBox changes, Log init moved to StandardagentpolisModule, fixed bug when test expected ProviderException instead of ProvisionException, checkConnection(connection) method in congestion model  structure test, Trip can now be constructed by varargs nodes, Bounds now return int in it|s methods instead of double, HighwayLayer moved to AgentPolis

+++ src/main/java/cz/agents/agentpolis/simulator/visualization/visio/Bounds.java
@@ -20,19 +20,19 @@ public class Bounds {
         return maxNode;
     }
 
-    public double getMaxLatE6() {
+    public int getMaxLatE6() {
         return maxNode.getLatE6();
     }
 
-    public double getMinLatE6() {
+    public int getMinLatE6() {
         return minNode.getLatE6();
     }
 
-    public double getMinLonE6() {
+    public int getMinLonE6() {
         return minNode.getLonE6();
     }
 
-    public double getMaxLonE6() {
+    public int getMaxLonE6() {
         return maxNode.getLonE6();
     }
 }
